+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Preflights can be imported (#181)
 - `--dont-stop` command line option, to try to continue even on errors (#209)
 - PDF/SVG PCB Print: option to print all pages/single page (#236)
+- iBoM: Support for variants that change component fields (#242)
 
 ### Fixed
 - OAR computation (Report) (#225)